int64_t in sum_8_7 and gemm_7_6

+++ onnx/version_converter/adapters/gemm_7_6.h
@@ -14,16 +14,16 @@ class Gemm_7_6 final : public Adapter {
       const ArrayRef<Value*>& inputs = node->inputs();
       // Determine if C is broadcastable
       // Get M and N
-      int M = inputs[0]->sizes()[0].dim;
+      int64_t M = inputs[0]->sizes()[0].dim;
       if (node->hasAttribute(ktransA) && node->i(ktransA) != 0)
         M = inputs[0]->sizes()[1].dim;
-      int N = inputs[1]->sizes()[1].dim;
+      int64_t N = inputs[1]->sizes()[1].dim;
       if (node->hasAttribute(ktransA) && node->i(ktransA) != 0)
         N = inputs[0]->sizes()[0].dim;
       const auto& C_shape = inputs[2]->sizes();
-      int C_M = C_shape[0].dim;
+      int64_t C_M = C_shape[0].dim;
       if (C_shape.size() == 2) {
-        int C_N = C_shape[1].dim;
+        int64_t C_N = C_shape[1].dim;
         ONNX_ASSERTM((C_M == M || C_M == 1) && (C_N == N || C_N == 1),
             "C not unidirectionally broadcastable to (M, N)");
       } else {
