+++ onnx/version_converter/adapters/sum_8_7.h
@@ -19,7 +19,7 @@ class Sum_8_7 final : public Adapter {
         std::vector<Dimension> B_sizes = inputs[i]->sizes();
         std::string error = "Sum in OpSet Version 6 does not support broadcasting";
         ONNX_ASSERT(A_sizes.size() == B_sizes.size());
-        for (int j = 1; j < A_sizes.size(); j++) {
+        for (int j = 1; j < (int) A_sizes.size(); j++) {
           ONNX_ASSERT(A_sizes[j - 1].dim == B_sizes[j].dim);
         }
       }
